+++ web/src/lib/components/shared-components/map/map.svelte
@@ -6,7 +6,7 @@
   import Icon from '$lib/components/elements/icon.svelte';
   import { colorTheme, mapSettings } from '$lib/stores/preferences.store';
   import { getAssetThumbnailUrl, handlePromiseError } from '$lib/utils';
-  import { Theme } from '$lib/utils/theme';
+  import { Theme } from '$lib/constants';
   import { getServerConfig, type MapMarkerResponseDto } from '@immich/sdk';
   import mapboxRtlUrl from '@mapbox/mapbox-gl-rtl-text/mapbox-gl-rtl-text.min.js?url';
   import { mdiCog, mdiMap, mdiMapMarker } from '@mdi/js';
