+++ web/src/lib/components/shared-components/theme-button.svelte
@@ -2,7 +2,7 @@
   import { moonPath, moonViewBox, sunPath, sunViewBox } from '$lib/assets/svg-paths';
   import CircleIconButton, { type Padding } from '$lib/components/elements/buttons/circle-icon-button.svelte';
   import { colorTheme, handleToggleTheme } from '$lib/stores/preferences.store';
-  import { Theme } from '$lib/utils/theme';
+  import { Theme } from '$lib/constants';
   import { t } from 'svelte-i18n';
 
   $: icon = $colorTheme.value === Theme.LIGHT ? moonPath : sunPath;
