+++ web/src/lib/constants.ts
@@ -102,6 +102,12 @@ export const timeBeforeShowLoadingSpinner: number = 100;
 
 export const timeDebounceOnSearch: number = 300;
 
+// should be the same values as the ones in the app.ts
+export enum Theme {
+  LIGHT = 'light',
+  DARK = 'dark',
+}
+
 export const fallbackLocale = {
   code: 'en-US',
   name: 'English (US)',
