--- web/src/lib/utils/theme.ts
@@ -1,11 +1,0 @@
-export interface ThemeSetting {
-  value: Theme;
-  system: boolean;
-}
-
-export enum Theme {
-  LIGHT = 'light',
-  DARK = 'dark',
-}
-
-export const colorThemeKeyName = 'color-theme';
